refactor: modify path of pages components to pages folder
